requirements.txt
celery==5.3.6
Flask==3.0.2
Flask_PyMongo==2.3.0
pandas==2.0.3
pymongo==4.6.2
python-dotenv==1.0.1
tabula_py==2.9.0
gunicorn==21.2.0